feat(irs-api):[#470] Add support for SingleLevelUsageAsPlanned
