+++ irs-api/src/test/java/org/eclipse/tractusx/irs/aaswrapper/job/delegate/RelationshipDelegateTest.java
@@ -24,12 +24,20 @@
 package org.eclipse.tractusx.irs.aaswrapper.job.delegate;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.eclipse.tractusx.irs.util.TestMother.SINGLE_LEVEL_BOM_AS_BUILT_3_0_0;
+import static org.eclipse.tractusx.irs.util.TestMother.SINGLE_LEVEL_BOM_AS_BUILT_3_1_0;
+import static org.eclipse.tractusx.irs.util.TestMother.SINGLE_LEVEL_BOM_AS_PLANNED_3_0_0;
+import static org.eclipse.tractusx.irs.util.TestMother.SINGLE_LEVEL_BOM_AS_PLANNED_3_1_0;
+import static org.eclipse.tractusx.irs.util.TestMother.SINGLE_LEVEL_USAGE_AS_BUILT_2_0_0;
+import static org.eclipse.tractusx.irs.util.TestMother.SINGLE_LEVEL_USAGE_AS_BUILT_2_1_0;
+import static org.eclipse.tractusx.irs.util.TestMother.SINGLE_LEVEL_USAGE_AS_PLANNED_2_0_0;
+import static org.eclipse.tractusx.irs.util.TestMother.SINGLE_LEVEL_USAGE_AS_PLANNED_2_1_0;
 import static org.eclipse.tractusx.irs.util.TestMother.jobParameter;
+import static org.eclipse.tractusx.irs.util.TestMother.jobParameterDownwardAsPlanned;
 import static org.eclipse.tractusx.irs.util.TestMother.jobParameterUpward;
+import static org.eclipse.tractusx.irs.util.TestMother.jobParameterUpwardAsPlanned;
 import static org.eclipse.tractusx.irs.util.TestMother.shell;
 import static org.eclipse.tractusx.irs.util.TestMother.shellDescriptor;
-import static org.eclipse.tractusx.irs.util.TestMother.singleLevelBomAsBuiltAspectName;
-import static org.eclipse.tractusx.irs.util.TestMother.singleLevelUsageAsBuiltAspectName;
 import static org.eclipse.tractusx.irs.util.TestMother.submodelDescriptorWithDspEndpoint;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
@@ -42,9 +50,11 @@ import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.List;
 import java.util.Objects;
+import java.util.stream.Stream;
 
 import org.eclipse.tractusx.irs.aaswrapper.job.AASTransferProcess;
 import org.eclipse.tractusx.irs.aaswrapper.job.ItemContainer;
+import org.eclipse.tractusx.irs.component.JobParameter;
 import org.eclipse.tractusx.irs.component.PartChainIdentificationKey;
 import org.eclipse.tractusx.irs.component.enums.ProcessStep;
 import org.eclipse.tractusx.irs.edc.client.EdcSubmodelFacade;
@@ -54,6 +64,9 @@ import org.eclipse.tractusx.irs.edc.client.model.SubmodelDescriptor;
 import org.eclipse.tractusx.irs.registryclient.discovery.ConnectorEndpointsService;
 import org.eclipse.tractusx.irs.util.JsonUtil;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
 
 class RelationshipDelegateTest {
 
@@ -69,13 +82,14 @@ class RelationshipDelegateTest {
         // given
         final String payload = Files.readString(
                 Paths.get(Objects.requireNonNull(getClass().getResource("/singleLevelBomAsBuilt.json")).toURI()));
-        when(submodelFacade.getSubmodelPayload(anyString(), anyString(), anyString(), any())).thenReturn(new SubmodelDescriptor("cid", payload));
+        when(submodelFacade.getSubmodelPayload(anyString(), anyString(), anyString(), any())).thenReturn(
+                new SubmodelDescriptor("cid", payload));
         when(connectorEndpointsService.fetchConnectorEndpoints(any())).thenReturn(List.of("http://localhost"));
 
         final ItemContainer.ItemContainerBuilder itemContainerWithShell = ItemContainer.builder()
                                                                                        .shell(shell("", shellDescriptor(
                                                                                                List.of(submodelDescriptorWithDspEndpoint(
-                                                                                                       singleLevelBomAsBuiltAspectName,
+                                                                                                       SINGLE_LEVEL_BOM_AS_BUILT_3_0_0,
                                                                                                        "address")))));
         final AASTransferProcess aasTransferProcess = new AASTransferProcess();
 
@@ -96,13 +110,14 @@ class RelationshipDelegateTest {
         // given
         final String payload = Files.readString(
                 Paths.get(Objects.requireNonNull(getClass().getResource("/singleLevelUsageAsBuilt.json")).toURI()));
-        when(submodelFacade.getSubmodelPayload(anyString(), anyString(), anyString(), any())).thenReturn(new SubmodelDescriptor("cid", payload));
+        when(submodelFacade.getSubmodelPayload(anyString(), anyString(), anyString(), any())).thenReturn(
+                new SubmodelDescriptor("cid", payload));
         when(connectorEndpointsService.fetchConnectorEndpoints(any())).thenReturn(List.of("http://localhost"));
 
         final ItemContainer.ItemContainerBuilder itemContainerWithShell = ItemContainer.builder()
                                                                                        .shell(shell("", shellDescriptor(
                                                                                                List.of(submodelDescriptorWithDspEndpoint(
-                                                                                                       singleLevelUsageAsBuiltAspectName,
+                                                                                                       SINGLE_LEVEL_USAGE_AS_BUILT_2_0_0,
                                                                                                        "address")))));
         final AASTransferProcess aasTransferProcess = new AASTransferProcess();
 
@@ -119,11 +134,134 @@ class RelationshipDelegateTest {
     }
 
     @Test
+    void shouldFillItemContainerWithUpwardAsPlannedRelationshipAndAddChildIdsToProcess()
+            throws EdcClientException, URISyntaxException, IOException {
+        // given
+        final String payload = Files.readString(Paths.get(
+                Objects.requireNonNull(getClass().getResource("/relationships/singleLevelUsageAsPlanned-2.0.0.json"))
+                       .toURI()));
+        when(submodelFacade.getSubmodelPayload(anyString(), anyString(), anyString(), any())).thenReturn(
+                new SubmodelDescriptor("cid", payload));
+        when(connectorEndpointsService.fetchConnectorEndpoints(any())).thenReturn(List.of("http://localhost"));
+
+        final ItemContainer.ItemContainerBuilder itemContainerWithShell = ItemContainer.builder()
+                                                                                       .shell(shell("", shellDescriptor(
+                                                                                               List.of(submodelDescriptorWithDspEndpoint(
+                                                                                                       SINGLE_LEVEL_USAGE_AS_PLANNED_2_0_0,
+                                                                                                       "address")))));
+        final AASTransferProcess aasTransferProcess = new AASTransferProcess();
+
+        // when
+        final ItemContainer result = relationshipDelegate.process(itemContainerWithShell, jobParameterUpwardAsPlanned(),
+                aasTransferProcess, createKey());
+
+        // then
+        assertThat(result).isNotNull();
+        assertThat(result.getRelationships()).isNotEmpty();
+        assertThat(aasTransferProcess.getIdsToProcess()).isNotEmpty();
+        assertThat(aasTransferProcess.getIdsToProcess().get(0).getGlobalAssetId()).isNotEmpty();
+        assertThat(aasTransferProcess.getIdsToProcess().get(0).getBpn()).isNotEmpty();
+    }
+
+    @ParameterizedTest
+    @MethodSource("relationshipParameters")
+    void shouldFillItemContainerWithSupportedRelationshipAndAddChildIdsToProcess(final String relationshipFile,
+            final String aspectName, final JobParameter jobParameter) throws Exception {
+        // given
+        final String payload = Files.readString(Paths.get(
+                Objects.requireNonNull(getClass().getResource("/relationships/" + relationshipFile)).toURI()));
+        when(submodelFacade.getSubmodelPayload(anyString(), anyString(), anyString(), any())).thenReturn(
+                new SubmodelDescriptor("cid", payload));
+        when(connectorEndpointsService.fetchConnectorEndpoints(any())).thenReturn(List.of("http://localhost"));
+
+        final ItemContainer.ItemContainerBuilder itemContainerWithShell = ItemContainer.builder()
+                                                                                       .shell(shell("", shellDescriptor(
+                                                                                               List.of(submodelDescriptorWithDspEndpoint(
+                                                                                                       aspectName,
+                                                                                                       "address")))));
+        final AASTransferProcess aasTransferProcess = new AASTransferProcess();
+
+        // when
+        final ItemContainer result = relationshipDelegate.process(itemContainerWithShell, jobParameter,
+                aasTransferProcess, createKey());
+
+        // then
+        assertThat(result).isNotNull();
+        assertThat(result.getRelationships()).isNotEmpty();
+        assertThat(aasTransferProcess.getIdsToProcess()).isNotEmpty();
+        assertThat(aasTransferProcess.getIdsToProcess().get(0).getGlobalAssetId()).isNotEmpty();
+        assertThat(aasTransferProcess.getIdsToProcess().get(0).getBpn()).isNotEmpty();
+    }
+
+    private static Stream<Arguments> relationshipParameters() {
+        return Stream.of(Arguments.of("singleLevelUsageAsPlanned-2.0.0.json", SINGLE_LEVEL_USAGE_AS_PLANNED_2_0_0,
+                        jobParameterUpwardAsPlanned()),
+                Arguments.of("singleLevelUsageAsBuilt-3.0.0.json", SINGLE_LEVEL_USAGE_AS_BUILT_2_0_0,
+                        jobParameterUpward()),
+                Arguments.of("singleLevelBomAsBuilt-3.0.0.json", SINGLE_LEVEL_BOM_AS_BUILT_3_0_0, jobParameter()),
+                Arguments.of("singleLevelBomAsPlanned-3.0.0.json", SINGLE_LEVEL_BOM_AS_PLANNED_3_0_0,
+                        jobParameterDownwardAsPlanned())
+                // asSpecified currently has no field for BPN of the child part
+                // Arguments.of("singleLevelBomAsSpecified-2.0.0.json", SINGLE_LEVEL_BOM_AS_SPECIFIED_2_0_0,
+                //         jobParameterDownwardAsSpecified()),
+                // Arguments.of("singleLevelBomAsSpecified-1.0.0.json", SINGLE_LEVEL_BOM_AS_SPECIFIED_1_0_0,
+                //         jobParameterDownwardAsSpecified())
+        );
+    }
+
+    @ParameterizedTest
+    @MethodSource("relationshipParametersFutureVersions")
+    void shouldFillItemContainerWithPotentialFutureMinorVersions(final String relationshipFile, final String aspectName,
+            final JobParameter jobParameter) throws Exception {
+        // given
+        final String payload = Files.readString(Paths.get(
+                Objects.requireNonNull(getClass().getResource("/relationships/futureVersions/" + relationshipFile))
+                       .toURI()));
+        when(submodelFacade.getSubmodelPayload(anyString(), anyString(), anyString(), any())).thenReturn(
+                new SubmodelDescriptor("cid", payload));
+        when(connectorEndpointsService.fetchConnectorEndpoints(any())).thenReturn(List.of("http://localhost"));
+
+        final ItemContainer.ItemContainerBuilder itemContainerWithShell = ItemContainer.builder()
+                                                                                       .shell(shell("", shellDescriptor(
+                                                                                               List.of(submodelDescriptorWithDspEndpoint(
+                                                                                                       aspectName,
+                                                                                                       "address")))));
+        final AASTransferProcess aasTransferProcess = new AASTransferProcess();
+
+        // when
+        final ItemContainer result = relationshipDelegate.process(itemContainerWithShell, jobParameter,
+                aasTransferProcess, createKey());
+
+        // then
+        assertThat(result).isNotNull();
+        assertThat(result.getRelationships()).isNotEmpty();
+        assertThat(aasTransferProcess.getIdsToProcess()).isNotEmpty();
+        assertThat(aasTransferProcess.getIdsToProcess().get(0).getGlobalAssetId()).isNotEmpty();
+        assertThat(aasTransferProcess.getIdsToProcess().get(0).getBpn()).isNotEmpty();
+    }
+
+    public static Stream<Arguments> relationshipParametersFutureVersions() {
+        return Stream.of(Arguments.of("singleLevelUsageAsPlanned-2.1.0.json", SINGLE_LEVEL_USAGE_AS_PLANNED_2_1_0,
+                        jobParameterUpwardAsPlanned()),
+                Arguments.of("singleLevelUsageAsBuilt-3.1.0.json", SINGLE_LEVEL_USAGE_AS_BUILT_2_1_0,
+                        jobParameterUpward()),
+                Arguments.of("singleLevelBomAsBuilt-3.1.0.json", SINGLE_LEVEL_BOM_AS_BUILT_3_1_0, jobParameter()),
+                Arguments.of("singleLevelBomAsPlanned-3.1.0.json", SINGLE_LEVEL_BOM_AS_PLANNED_3_1_0,
+                        jobParameterDownwardAsPlanned())
+                // asSpecified currently has no field for BPN of the child part
+                // Arguments.of("singleLevelBomAsSpecified-2.1.0.json", SINGLE_LEVEL_BOM_AS_SPECIFIED_2_1_0,
+                //         jobParameterDownwardAsSpecified()),
+                // Arguments.of("singleLevelBomAsSpecified-1.1.0.json", SINGLE_LEVEL_BOM_AS_SPECIFIED_1_1_0,
+                //         jobParameterDownwardAsSpecified())
+        );
+    }
+
+    @Test
     void shouldPutTombstoneForMissingBpn() {
         final ItemContainer.ItemContainerBuilder itemContainerWithShell = ItemContainer.builder()
                                                                                        .shell(shell("", shellDescriptor(
                                                                                                List.of(submodelDescriptorWithDspEndpoint(
-                                                                                                       singleLevelBomAsBuiltAspectName,
+                                                                                                       SINGLE_LEVEL_BOM_AS_BUILT_3_0_0,
                                                                                                        "address")))));
         // when
         final ItemContainer result = relationshipDelegate.process(itemContainerWithShell, jobParameter(),
@@ -147,7 +285,7 @@ class RelationshipDelegateTest {
         final ItemContainer.ItemContainerBuilder itemContainerWithShell = ItemContainer.builder()
                                                                                        .shell(shell("", shellDescriptor(
                                                                                                List.of(submodelDescriptorWithDspEndpoint(
-                                                                                                       singleLevelBomAsBuiltAspectName,
+                                                                                                       SINGLE_LEVEL_BOM_AS_BUILT_3_0_0,
                                                                                                        "address")))));
 
         // when
@@ -171,7 +309,7 @@ class RelationshipDelegateTest {
         final ItemContainer.ItemContainerBuilder itemContainerWithShell = ItemContainer.builder()
                                                                                        .shell(shell("", shellDescriptor(
                                                                                                List.of(submodelDescriptorWithDspEndpoint(
-                                                                                                       singleLevelBomAsBuiltAspectName,
+                                                                                                       SINGLE_LEVEL_BOM_AS_BUILT_3_0_0,
                                                                                                        "address")))));
 
         // when
@@ -193,11 +331,12 @@ class RelationshipDelegateTest {
         final ItemContainer.ItemContainerBuilder itemContainerWithShell = ItemContainer.builder()
                                                                                        .shell(shell("", shellDescriptor(
                                                                                                List.of(submodelDescriptorWithDspEndpoint(
-                                                                                                       singleLevelBomAsBuiltAspectName,
+                                                                                                       SINGLE_LEVEL_BOM_AS_BUILT_3_0_0,
                                                                                                        "address")))));
 
         // when
-        when(submodelFacade.getSubmodelPayload(any(), any(), any(), any())).thenThrow(new UsagePolicyException("itemId", null, businessPartnerNumber));
+        when(submodelFacade.getSubmodelPayload(any(), any(), any(), any())).thenThrow(
+                new UsagePolicyException("itemId", null, businessPartnerNumber));
         when(connectorEndpointsService.fetchConnectorEndpoints(any())).thenReturn(List.of("connector.endpoint.nl"));
         final ItemContainer result = relationshipDelegate.process(itemContainerWithShell, jobParameter(),
                 new AASTransferProcess(), createKey());
