+++ irs-api/src/test/resources/relationships/futureVersions/singleLevelBomAsPlanned-3.1.0.json
@@ -1,0 +1,20 @@
+{
+  "catenaXId": "urn:uuid:055c1128-0375-47c8-98de-7cf802c3241d",
+  "childItems": [
+    {
+      "validityPeriod": {
+        "validFrom": "2023-03-21T08:17:29.187+01:00",
+        "validTo": "2024-07-01T16:10:00.000+01:00"
+      },
+      "catenaXId": "urn:uuid:cc6a703e-ab0e-417c-9605-c9c70e33a47c",
+      "quantity": {
+        "value": 20.0,
+        "unit": "unit:piece"
+      },
+      "createdOn": "2022-02-03T14:48:54.709Z",
+      "businessPartner": "BPNL50096894aNXY",
+      "lastModifiedOn": "2022-02-03T14:48:54.709Z",
+      "futureAddedField": "test"
+    }
+  ]
+}
