+++ irs-api/src/test/resources/relationships/futureVersions/singleLevelBomAsSpecified-2.1.0.json
@@ -1,0 +1,31 @@
+{
+  "manufacturerId": "BPNL1234567890LR",
+  "assetId": "urn:uuid:055c1128-0375-47c8-98de-7cf802c3241d",
+  "childItems": [
+    {
+      "childItemCategory": "e.g. vehicle, winter wheels, bicycle rack",
+      "item": [
+        {
+          "itemClassification": [
+            {
+              "value": "STEEWHL",
+              "key": "BMW:PartFamily"
+            }
+          ],
+          "itemQuantity": {
+            "quantityNumber": 350.0,
+            "measurementUnit": "unit:kW"
+          },
+          "ownerItemId": "urn:uuid:b4741433-92bb-4027-9a02-bbc64a58d193",
+          "itemVersion": "05",
+          "itemDescription": "The steering wheel is nice and round",
+          "createdOn": "2022-02-03T14:48:54.709Z",
+          "itemPositioning": "left",
+          "lastModifiedOn": "2022-02-03T14:48:54.709Z"
+        }
+      ],
+      "childassetId": "urn:uuid:218b26f4-4a0b-4a7f-b2c1-d248927718bf",
+      "futureAddedField": "test"
+    }
+  ]
+}
