+++ irs-api/src/test/resources/relationships/singleLevelBomAsSpecified-1.0.0.json
@@ -1,0 +1,28 @@
+{
+  "catenaXId": "urn:uuid:055c1128-0375-47c8-98de-7cf802c3241d",
+  "childParts": [
+    {
+      "childPartsCategory": "e.g. vehicle, winter wheels, bicycle rack",
+      "part": [
+        {
+          "ownerPartId": "uuid!",
+          "partVersion": "05",
+          "partQuantity": {
+            "quantityNumber": 350.0,
+            "measurementUnit": "kW"
+          },
+          "partDescription": "The steering wheel is nice and round",
+          "partClassification": [
+            {
+              "value": "STEEWHL",
+              "key": "BMW:PartFamily"
+            }
+          ],
+          "createdOn": "2022-02-03T14:48:54.709Z",
+          "lastModifiedOn": "2022-02-03T14:48:54.709Z"
+        }
+      ],
+      "childCatenaXId": "urn:uuid:218b26f4-4a0b-4a7f-b2c1-d248927718bf "
+    }
+  ]
+}
