+++ irs-api/src/test/resources/relationships/singleLevelUsageAsBuilt-3.0.0.json
@@ -1,0 +1,19 @@
+{
+  "catenaXId": "urn:uuid:055c1128-0375-47c8-98de-7cf802c3241d",
+  "parentItems": [
+    {
+      "catenaXId": "urn:uuid:cc6a703e-ab0e-417c-9605-c9c70e33a47c",
+      "isOnlyPotentialParent": false,
+      "quantity": {
+        "value": 20.0,
+        "unit": "unit:piece"
+      },
+      "createdOn": "2022-02-03T14:48:54.709Z",
+      "businessPartner": "BPNL50096894aNXY",
+      "lastModifiedOn": "2022-02-03T14:48:54.709Z"
+    }
+  ],
+  "customers": [
+    "BPNL50096894aNXY"
+  ]
+}
