+++ irs-api/src/test/resources/relationships/singleLevelUsageAsPlanned-2.0.0.json
@@ -1,0 +1,22 @@
+{
+  "catenaXId": "urn:uuid:055c1128-0375-47c8-98de-7cf802c3241d",
+  "parentItems": [
+    {
+      "validityPeriod": {
+        "validFrom": "2023-03-21T08:17:29.187+01:00",
+        "validTo": "2024-07-01T16:10:00.000+01:00"
+      },
+      "catenaXId": "urn:uuid:56319907-28dc-440e-afcc-72d67ad343e7",
+      "quantity": {
+        "value": 20.0,
+        "unit": "unit:piece"
+      },
+      "createdOn": "2022-02-03T14:48:54.709Z",
+      "businessPartner": "BPNL50096894aNXY",
+      "lastModifiedOn": "2022-02-03T14:48:54.709Z"
+    }
+  ],
+  "customers": [
+    "BPNL50096894aNXY"
+  ]
+}
