+++ irs-edc-client/src/main/java/org/eclipse/tractusx/irs/edc/client/relationships/SingleLevelUsageAsBuilt.java
@@ -1,0 +1,65 @@
+/********************************************************************************
+ * Copyright (c) 2022,2024
+ *       2022: ZF Friedrichshafen AG
+ *       2022: ISTOS GmbH
+ *       2022,2024: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *       2022,2023: BOSCH AG
+ * Copyright (c) 2021,2024 Contributors to the Eclipse Foundation
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Apache License, Version 2.0 which is available at
+ * https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ********************************************************************************/
+package org.eclipse.tractusx.irs.edc.client.relationships;
+
+import java.util.Collection;
+import java.util.List;
+import java.util.Optional;
+import java.util.Set;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.extern.jackson.Jacksonized;
+import org.eclipse.tractusx.irs.component.Relationship;
+import org.eclipse.tractusx.irs.component.enums.AspectType;
+import org.eclipse.tractusx.irs.component.enums.BomLifecycle;
+import org.eclipse.tractusx.irs.edc.client.RelationshipSubmodel;
+
+/**
+ * SingleLevelUsageAsBuilt
+ */
+@Data
+@Builder
+@Jacksonized
+@AllArgsConstructor
+@NoArgsConstructor
+class SingleLevelUsageAsBuilt implements RelationshipSubmodel {
+
+    private String catenaXId;
+    private Set<ParentItem> parentItems;
+    private Set<String> customers;
+
+    @Override
+    public List<Relationship> asRelationships() {
+        return Optional.ofNullable(this.parentItems)
+                       .stream()
+                       .flatMap(Collection::stream)
+                       .map(parentItem -> parentItem.toRelationship(this.catenaXId, BomLifecycle.AS_BUILT,
+                               AspectType.SINGLE_LEVEL_USAGE_AS_BUILT))
+                       .toList();
+    }
+
+}
